requirements.txt
Flask==1.0.3
google-cloud-storage==1.16.0
google-api-python-client==1.7.9
